refactor: split out create sequence statement and sequence def
